requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
./libs/TTS-0.8.0-py3-none-any.whl
pydub==0.25.1
tqdm==4.66.1
python-multipart==0.0.6 